+++ packages/plugin-vue/src/index.js
@@ -34,3 +34,6 @@ module.exports = class BugsnagPluginVue {
     }
   }
 }
+
+// add a default export for ESM modules without interop
+module.exports.default = module.exports